PdfExtract/requirements.txt

Django==3.2
django-cors-headers==3.7.0
pdfminer.six==20201018
python-docx==0.8.11